Spark_Task/requirements.txt
# Spark ETL Dependencies
pyspark==3.5.0
python-geohash==0.8.5
requests==2.31.0
pytest==7.4.3
pytest-cov==4.1.0
pandas==2.1.4
pyarrow==14.0.2
